Update Calculate_theta.js
You can copy paste this into a chrome browser after using ctrl+shift+i, then going to the console and paste the code in. From there select the function, input the numbers(o=opposite, h=hypotenuse, a=adjacent), and run it. Note: This version is for JS only; not for an html page. The other JS page(Calculate-angles.js) is the on that goes with the HTML and CSS code.